Chore: Refactor the codebase (fixes #220)
