+++ .travis.yml
@@ -1,8 +1,9 @@
 language: node_js
 sudo: false
 node_js:
-    - "0.12"
-    - 4
-    - 6
+    - "4"
+    - "5"
+    - "6"
+    - "7"
 after_success:
     - npm run coveralls
