+++ lib/ast-node-types.js
@@ -18,37 +18,60 @@
 //------------------------------------------------------------------------------
 
 module.exports = {
-    AssignmentExpression: "AssignmentExpression",
-    AssignmentPattern: "AssignmentPattern",
     ArrayExpression: "ArrayExpression",
     ArrayPattern: "ArrayPattern",
     ArrowFunctionExpression: "ArrowFunctionExpression",
+    AssignmentExpression: "AssignmentExpression",
+    AssignmentPattern: "AssignmentPattern",
     AwaitExpression: "AwaitExpression",
-    BlockStatement: "BlockStatement",
     BinaryExpression: "BinaryExpression",
+    BlockStatement: "BlockStatement",
     BreakStatement: "BreakStatement",
     CallExpression: "CallExpression",
     CatchClause: "CatchClause",
     ClassBody: "ClassBody",
     ClassDeclaration: "ClassDeclaration",
     ClassExpression: "ClassExpression",
+    ClassImplements: "ClassImplements",
+    ClassProperty: "ClassProperty",
     ConditionalExpression: "ConditionalExpression",
     ContinueStatement: "ContinueStatement",
-    DoWhileStatement: "DoWhileStatement",
     DebuggerStatement: "DebuggerStatement",
+    DeclareFunction: "DeclareFunction",
+    DoWhileStatement: "DoWhileStatement",
     EmptyStatement: "EmptyStatement",
     ExperimentalRestProperty: "ExperimentalRestProperty",
     ExperimentalSpreadProperty: "ExperimentalSpreadProperty",
+    ExportAllDeclaration: "ExportAllDeclaration",
+    ExportDefaultDeclaration: "ExportDefaultDeclaration",
+    ExportNamedDeclaration: "ExportNamedDeclaration",
+    ExportSpecifier: "ExportSpecifier",
     ExpressionStatement: "ExpressionStatement",
-    ForStatement: "ForStatement",
     ForInStatement: "ForInStatement",
     ForOfStatement: "ForOfStatement",
+    ForStatement: "ForStatement",
     FunctionDeclaration: "FunctionDeclaration",
     FunctionExpression: "FunctionExpression",
+    GenericTypeAnnotation: "GenericTypeAnnotation",
     Identifier: "Identifier",
     IfStatement: "IfStatement",
-    Literal: "Literal",
+    ImportDeclaration: "ImportDeclaration",
+    ImportDefaultSpecifier: "ImportDefaultSpecifier",
+    ImportNamespaceSpecifier: "ImportNamespaceSpecifier",
+    ImportSpecifier: "ImportSpecifier",
+    JSXAttribute: "JSXAttribute",
+    JSXClosingElement: "JSXClosingElement",
+    JSXElement: "JSXElement",
+    JSXEmptyExpression: "JSXEmptyExpression",
+    JSXExpressionContainer: "JSXExpressionContainer",
+    JSXIdentifier: "JSXIdentifier",
+    JSXMemberExpression: "JSXMemberExpression",
+    JSXNamespacedName: "JSXNamespacedName",
+    JSXOpeningElement: "JSXOpeningElement",
+    JSXSpreadAttribute: "JSXSpreadAttribute",
+    JSXText: "JSXText",
     LabeledStatement: "LabeledStatement",
+    Literal: "Literal",
     LogicalExpression: "LogicalExpression",
     MemberExpression: "MemberExpression",
     MetaProperty: "MetaProperty",
@@ -71,30 +94,46 @@ module.exports = {
     ThisExpression: "ThisExpression",
     ThrowStatement: "ThrowStatement",
     TryStatement: "TryStatement",
+
+    /**
+     * TS-prefixed nodes
+     */
+    TSAbstractClassProperty: "TSAbstractClassProperty",
+    TSAbstractMethodDefinition: "TSAbstractMethodDefinition",
+    TSAnyKeyword: "TSAnyKeyword",
+    TSArrayType: "TSArrayType",
+    TSBooleanKeyword: "TSBooleanKeyword",
+    TSConstructorType: "TSConstructorType",
+    TSConstructSignature: "TSConstructSignature",
+    TSDeclareKeyword: "TSDeclareKeyword",
+    TSInterfaceBody: "TSInterfaceBody",
+    TSInterfaceDeclaration: "TSInterfaceDeclaration",
+    TSInterfaceHeritage: "TSInterfaceHeritage",
+    TSFunctionType: "TSFunctionType",
+    TSMethodSignature: "TSMethodSignature",
+    TSModuleBlock: "TSModuleBlock",
+    TSModuleDeclaration: "TSModuleDeclaration",
+    TSNamespaceFunctionDeclaration: "TSNamespaceFunctionDeclaration",
+    TSNonNullExpression: "TSNonNullExpression",
+    TSNumberKeyword: "TSNumberKeyword",
+    TSParameterProperty: "TSParameterProperty",
+    TSPropertySignature: "TSPropertySignature",
+    TSQuestionToken: "TSQuestionToken",
+    TSStringKeyword: "TSStringKeyword",
+    TSTypeLiteral: "TSTypeLiteral",
+    TSTypeReference: "TSTypeReference",
+    TSUnionType: "TSUnionType",
+    TSVoidKeyword: "TSVoidKeyword",
+
+    TypeAnnotation: "TypeAnnotation",
+    TypeParameter: "TypeParameter",
+    TypeParameterDeclaration: "TypeParameterDeclaration",
+    TypeParameterInstantiation: "TypeParameterInstantiation",
     UnaryExpression: "UnaryExpression",
     UpdateExpression: "UpdateExpression",
     VariableDeclaration: "VariableDeclaration",
     VariableDeclarator: "VariableDeclarator",
     WhileStatement: "WhileStatement",
     WithStatement: "WithStatement",
-    YieldExpression: "YieldExpression",
-    JSXIdentifier: "JSXIdentifier",
-    JSXNamespacedName: "JSXNamespacedName",
-    JSXMemberExpression: "JSXMemberExpression",
-    JSXEmptyExpression: "JSXEmptyExpression",
-    JSXExpressionContainer: "JSXExpressionContainer",
-    JSXElement: "JSXElement",
-    JSXClosingElement: "JSXClosingElement",
-    JSXOpeningElement: "JSXOpeningElement",
-    JSXAttribute: "JSXAttribute",
-    JSXSpreadAttribute: "JSXSpreadAttribute",
-    JSXText: "JSXText",
-    ExportDefaultDeclaration: "ExportDefaultDeclaration",
-    ExportNamedDeclaration: "ExportNamedDeclaration",
-    ExportAllDeclaration: "ExportAllDeclaration",
-    ExportSpecifier: "ExportSpecifier",
-    ImportDeclaration: "ImportDeclaration",
-    ImportSpecifier: "ImportSpecifier",
-    ImportDefaultSpecifier: "ImportDefaultSpecifier",
-    ImportNamespaceSpecifier: "ImportNamespaceSpecifier"
+    YieldExpression: "YieldExpression"
 };
