--- lib/features.js
@@ -1,94 +1,0 @@
-/**
- * @fileoverview The list of feature flags supported by the parser and their default
- *      settings.
- * @author Nicholas C. Zakas
- * @copyright jQuery Foundation and other contributors, https://jquery.org/
- * MIT License
- */
-
-"use strict";
-
-//------------------------------------------------------------------------------
-// Requirements
-//------------------------------------------------------------------------------
-
-// None!
-
-//------------------------------------------------------------------------------
-// Public
-//------------------------------------------------------------------------------
-
-module.exports = {
-
-    // enable parsing of arrow functions
-    arrowFunctions: false,
-
-    // enable parsing of let and const
-    blockBindings: false,
-
-    // enable parsing of destructured arrays and objects
-    destructuring: false,
-
-    // enable parsing of regex u flag
-    regexUFlag: false,
-
-    // enable parsing of regex y flag
-    regexYFlag: false,
-
-    // enable parsing of template strings
-    templateStrings: false,
-
-    // enable parsing binary literals
-    binaryLiterals: false,
-
-    // enable parsing ES6 octal literals
-    octalLiterals: false,
-
-    // enable parsing unicode code point escape sequences
-    unicodeCodePointEscapes: true,
-
-    // enable parsing of default parameters
-    defaultParams: false,
-
-    // enable parsing of rest parameters
-    restParams: false,
-
-    // enable parsing of for-of statements
-    forOf: false,
-
-    // enable parsing computed object literal properties
-    objectLiteralComputedProperties: false,
-
-    // enable parsing of shorthand object literal methods
-    objectLiteralShorthandMethods: false,
-
-    // enable parsing of shorthand object literal properties
-    objectLiteralShorthandProperties: false,
-
-    // Allow duplicate object literal properties (except '__proto__')
-    objectLiteralDuplicateProperties: false,
-
-    // enable parsing of generators/yield
-    generators: false,
-
-    // support the spread operator
-    spread: false,
-
-    // enable parsing of classes
-    classes: false,
-
-    // enable parsing of new.target
-    newTarget: false,
-
-    // enable parsing of modules
-    modules: false,
-
-    // React JSX parsing
-    jsx: false,
-
-    // allow return statement in global scope
-    globalReturn: false,
-
-    // allow experimental object rest/spread
-    experimentalObjectRestSpread: false
-};
