+++ tests/lib/basics.js
@@ -11,7 +11,7 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-var assert = require("chai").assert,
+const assert = require("chai").assert,
     leche = require("leche"),
     path = require("path"),
     parser = require("../../parser"),
@@ -22,39 +22,40 @@ var assert = require("chai").assert,
 // Setup
 //------------------------------------------------------------------------------
 
-var FIXTURES_DIR = "./tests/fixtures/basics";
+const FIXTURES_DIR = "./tests/fixtures/basics";
 
-var testFiles = shelljs.find(FIXTURES_DIR).filter(function(filename) {
-    return filename.indexOf(".src.js") > -1;
-}).map(function(filename) {
-    return filename.substring(FIXTURES_DIR.length - 1, filename.length - 7);  // strip off ".src.js"
-});
+const testFiles = shelljs.find(FIXTURES_DIR)
+    .filter(filename => filename.indexOf(".src.js") > -1)
+    // strip off ".src.js"
+    .map(filename => filename.substring(FIXTURES_DIR.length - 1, filename.length - 7));
 
 //------------------------------------------------------------------------------
 // Tests
 //------------------------------------------------------------------------------
 
-describe("basics", function() {
+describe("basics", () => {
 
-    var config;
+    let config;
 
-    beforeEach(function() {
+    beforeEach(() => {
         config = {
             loc: true,
             range: true,
             tokens: true,
-            ecmaFeatures: {}
+            ecmaFeatures: {},
+            errorOnUnknownASTType: true
         };
     });
 
-    leche.withData(testFiles, function(filename) {
+    leche.withData(testFiles, filename => {
+
         // Uncomment and fill in filename to focus on a single file
         // var filename = "jsx/invalid-matching-placeholder-in-closing-tag";
-        var code = shelljs.cat(path.resolve(FIXTURES_DIR, filename) + ".src.js");
+        const code = shelljs.cat(`${path.resolve(FIXTURES_DIR, filename)}.src.js`);
 
-        it("should parse correctly", function() {
-            var expected = require(path.resolve(__dirname, "../../", FIXTURES_DIR, filename) + ".result.js");
-            var result;
+        it("should parse correctly", () => {
+            const expected = require(`${path.resolve(__dirname, "../../", FIXTURES_DIR, filename)}.result.js`);
+            let result;
 
             try {
                 result = parser.parse(code, config);
@@ -64,11 +65,12 @@ describe("basics", function() {
                 // format of error isn't exactly the same, just check if it's expected
                 if (expected.message) {
                     return;
-                } else {
-                    throw ex;
                 }
+                throw ex;
+
 
             }
+
             // console.log(JSON.stringify(result, null, 4));
             assert.deepEqual(result, expected);
